@@ -31,12 +31,18 @@
         "signature_pad": "^2.3.2",
         "simpl-schema": "^0.3.2",
         "square-connect": "^2.20180918.1",
-        "string": "^3.3.3"
+        "string": "^3.3.3",
+        "typescript": "^3.3.0-dev.20190115"
     },
     "devDependencies": {
         "babel-cli": "^6.26.0",
+        "babel-core": "^6.26.3",
+        "babel-eslint": "^10.0.1",
+        "babel-preset-es2015": "^6.24.1",
         "babel-preset-flow": "^6.23.0",
-        "eslint": "^5.11.0",
+        "babel-preset-react": "^6.24.1",
+        "babel-preset-stage-0": "^6.24.1",
+        "eslint": "^5.12.0",
         "eslint-config-standard": "^10.2.1",
         "eslint-plugin-import": "^2.14.0",
         "eslint-plugin-node": "^5.2.1",
